Finally got composite key right

+++ model/src/main/scala/gr/grnet/aquarium/model/ServiceItemConfig.scala
@@ -2,24 +2,22 @@ package gr.grnet.aquarium.model
 
 import javax.persistence._
 
-case class ServiceItemConfigKey(var item_id : Int, var resource_id : Int)
+case class ServiceItemConfigKey(var item_id : Long, var resource_id : Long)
 
 @Table(name = "SERVICE_ITEM_CONFIG")
 @IdClass(classOf[ServiceItemConfigKey])
 @javax.persistence.Entity
 class ServiceItemConfig {
 
-  //def this() = this("")
-
   @javax.persistence.Id
   @Column(name ="SERVICE_ITEM_ID", nullable=false, updatable=false,
           insertable=false)
-  var item_id: Int = _
+  var item_id: Long = _
 
   @javax.persistence.Id
   @Column(name ="CONSUMABLE_RESOURCE_ID", nullable=false, updatable=false,
           insertable=false)
-  var resource_id: Int = _
+  var resource_id: Long = _
 
   @ManyToOne (cascade = Array(CascadeType.ALL),
               targetEntity = classOf[ServiceItem])
